Restructured the code.
